+++ src/main/java/com/yanny/ytech/compatibility/emi/EmiWorkspaceCraftingRecipe.java
@@ -1,7 +1,6 @@
 package com.yanny.ytech.compatibility.emi;
 
 import com.yanny.ytech.configuration.recipe.WorkspaceCraftingRecipe;
-import com.yanny.ytech.registration.YTechItemTags;
 import com.yanny.ytech.registration.YTechItems;
 import com.yanny.ytech.registration.YTechRecipeTypes;
 import dev.emi.emi.api.recipe.BasicEmiRecipe;
@@ -10,7 +9,6 @@ import dev.emi.emi.api.render.EmiTexture;
 import dev.emi.emi.api.stack.EmiIngredient;
 import dev.emi.emi.api.stack.EmiStack;
 import dev.emi.emi.api.widget.WidgetHolder;
-import net.minecraft.world.item.crafting.Ingredient;
 
 import java.util.List;
 
@@ -25,7 +23,7 @@ public class EmiWorkspaceCraftingRecipe extends BasicEmiRecipe {
         super(CATEGORY, recipe.getId(), 154, 164);
         id = recipe.getId();
         inputs = recipe.recipeItems().stream().map(EmiIngredient::of).toList();
-        catalysts = List.of(EmiIngredient.of(Ingredient.of(YTechItemTags.SHARP_FLINTS)));
+        catalysts = List.of(EmiIngredient.of(recipe.tool()));
         outputs = List.of(EmiStack.of(recipe.result()));
     }