+++ src/main/java/com/yanny/ytech/configuration/block/PrimitiveAlloySmelterBlock.java
@@ -4,6 +4,7 @@ import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.block_entity.PrimitiveAlloySmelterBlockEntity;
 import com.yanny.ytech.configuration.recipe.WorkspaceCraftingRecipe;
 import com.yanny.ytech.registration.YTechBlocks;
+import com.yanny.ytech.registration.YTechItemTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.data.recipes.FinishedRecipe;
@@ -51,7 +52,7 @@ public class PrimitiveAlloySmelterBlock extends AbstractPrimitiveMachineBlock {
     }
 
     public static void registerRecipe(@NotNull Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(YTechBlocks.PRIMITIVE_ALLOY_SMELTER.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, YTechBlocks.PRIMITIVE_ALLOY_SMELTER.get())
                 .define('A', Items.FURNACE)
                 .define('C', Items.BRICKS)
                 .bottomPattern("CCC")