+++ src/main/java/com/yanny/ytech/generation/YTechRecipeProvider.java
@@ -1678,7 +1678,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void registerStrainerRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(YTechItems.STRAINER.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.SHARP_FLINTS, YTechItems.STRAINER.get())
                 .define('L', ItemTags.LOGS)
                 .define('W', YTechItemTags.WOODEN_BOXES)
                 .bottomPattern("LLL")
@@ -1756,7 +1756,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void registerTreeStumpRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(YTechItems.TREE_STUMP.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.AXES.tag, YTechItems.TREE_STUMP.get())
                 .define('C', ItemTags.LOGS)
                 .define('S', Tags.Items.COBBLESTONE_NORMAL)
                 .bottomPattern("SSS")
@@ -1793,7 +1793,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcChestRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.CHEST)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.CHEST)
                 .define('C', ItemTags.PLANKS)
                 .define('R', YTechItemTags.RODS.get(IRON))
                 .define('B', ItemTags.WOODEN_BUTTONS)
@@ -1811,7 +1811,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcFurnaceRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.FURNACE)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.FURNACE)
                 .define('C', Tags.Items.COBBLESTONE_NORMAL)
                 .define('A', Items.CAMPFIRE)
                 .bottomPattern("CCC")
@@ -1828,7 +1828,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcCraftingTableRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.CRAFTING_TABLE)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.CRAFTING_TABLE)
                 .define('L', Items.LEATHER)
                 .define('P', ItemTags.PLANKS)
                 .bottomPattern("PPP")
@@ -1845,7 +1845,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcStonecutterRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.STONECUTTER)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.STONECUTTER)
                 .define('L', ItemTags.LOGS)
                 .define('P', Items.SMOOTH_STONE)
                 .define('S', YTechItemTags.SAW_BLADES.tag)
@@ -1863,7 +1863,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcBarrelRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.BARREL)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.BARREL)
                 .define('S', ItemTags.WOODEN_SLABS)
                 .define('P', ItemTags.PLANKS)
                 .bottomPattern("PPP")
@@ -1880,7 +1880,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcFletchingTableRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.FLETCHING_TABLE)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.FLETCHING_TABLE)
                 .define('T', Items.TARGET)
                 .define('P', ItemTags.PLANKS)
                 .bottomPattern("PPP")
@@ -1897,7 +1897,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcCartographyTableRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.CARTOGRAPHY_TABLE)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.CARTOGRAPHY_TABLE)
                 .define('P', ItemTags.PLANKS)
                 .define('M', Items.MAP)
                 .bottomPattern("PPP")
@@ -1914,7 +1914,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcSmithingTableRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.SMITHING_TABLE)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.SMITHING_TABLE)
                 .define('P', ItemTags.PLANKS)
                 .define('C', YTechItemTags.STORAGE_BLOCKS.get(IRON))
                 .bottomPattern("CPC")
@@ -1931,7 +1931,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcGrindstoneRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.GRINDSTONE)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.GRINDSTONE)
                 .define('L', ItemTags.LOGS)
                 .define('S', Items.SMOOTH_STONE)
                 .bottomPattern(" L ")
@@ -1948,7 +1948,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcLoomRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.LOOM)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.LOOM)
                 .define('P', ItemTags.PLANKS)
                 .define('S', Tags.Items.STRING)
                 .bottomPattern("P P")
@@ -1965,7 +1965,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcSmokerRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.SMOKER)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.SMOKER)
                 .define('L', ItemTags.LOGS)
                 .define('A', Items.CAMPFIRE)
                 .define('C', Tags.Items.COBBLESTONE_NORMAL)
@@ -1983,7 +1983,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcComposterRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(Items.COMPOSTER)
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, Items.COMPOSTER)
                 .define('L', ItemTags.PLANKS)
                 .bottomPattern("LLL")
                 .bottomPattern("LLL")
@@ -1999,7 +1999,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcWoodenBoxRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(YTechItems.WOODEN_BOX.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, YTechItems.WOODEN_BOX.get())
                 .define('L', ItemTags.PLANKS)
                 .bottomPattern("LLL")
                 .bottomPattern("LLL")
@@ -2015,7 +2015,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcToolRackRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(YTechItems.TOOL_RACK.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, YTechItems.TOOL_RACK.get())
                 .define('L', ItemTags.PLANKS)
                 .define('B', YTechItemTags.BOLTS.tag)
                 .bottomPattern("LLL")
@@ -2032,7 +2032,7 @@ class YTechRecipeProvider extends RecipeProvider {
     }
 
     private static void wcWellPulleyRecipe(Consumer<FinishedRecipe> recipeConsumer) {
-        WorkspaceCraftingRecipe.Builder.recipe(YTechItems.WELL_PULLEY.get())
+        WorkspaceCraftingRecipe.Builder.recipe(YTechItemTags.HAMMERS.tag, YTechItems.WELL_PULLEY.get())
                 .define('L', ItemTags.LOGS)
                 .define('P', ItemTags.PLANKS)
                 .define('B', YTechItemTags.TERRACOTTA_BRICKS)